Revert "make header count ulong"
This reverts commit d5be2fa8c57696aaf090e8bf911fbc70033c089e.

+++ include/parse.h
@@ -9,7 +9,7 @@ struct dbheader_t
 {
     unsigned int magic;
     unsigned short version;
-    unsigned long count;
+    unsigned short count;
     unsigned int filesize;
 };
